Split FileUtils into FileReaderUtils and FileMutatorUtils to separate reads from mutations.
