+++ apps/desktop/desktop/src/main/java/bisq/desktop/common/standby/SoundPlayer.java
@@ -17,7 +17,7 @@
 
 package bisq.desktop.common.standby;
 
-import bisq.common.file.FileUtils;
+import bisq.common.file.FileMutatorUtils;
 import bisq.common.threading.ExecutorFactory;
 import bisq.desktop.ServiceProvider;
 import lombok.extern.slf4j.Slf4j;
@@ -70,7 +70,7 @@ class SoundPlayer implements PreventStandbyMode {
             String fileName = "prevent-app-nap-silent-sound.aiff";
             Path soundFilePath = appDataDirPath.resolve(fileName);
             if (!Files.exists(soundFilePath)) {
-                FileUtils.resourceToFile(fileName, soundFilePath);
+                FileMutatorUtils.resourceToFile(fileName, soundFilePath);
             }
             AudioInputStream audioInputStream = null;
             SourceDataLine sourceDataLine = null;
