+++ common/src/main/java/bisq/common/file/FileMutatorUtils.java
@@ -22,14 +22,12 @@ import bisq.common.observable.Observable;
 import lombok.extern.slf4j.Slf4j;
 
 import java.io.BufferedInputStream;
-import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.net.HttpURLConnection;
 import java.net.URISyntaxException;
 import java.net.URL;
-import java.net.URLDecoder;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.FileVisitResult;
 import java.nio.file.Files;
@@ -38,30 +36,12 @@ import java.nio.file.SimpleFileVisitor;
 import java.nio.file.StandardCopyOption;
 import java.nio.file.attribute.BasicFileAttributes;
 import java.util.Comparator;
-import java.util.Enumeration;
-import java.util.HashSet;
-import java.util.Optional;
-import java.util.Scanner;
 import java.util.Set;
-import java.util.concurrent.TimeoutException;
 import java.util.concurrent.atomic.AtomicReference;
-import java.util.jar.JarEntry;
-import java.util.jar.JarFile;
-import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 @Slf4j
-public class FileUtils {
-    public static final String FILE_SEP = File.separator;
-
-    public static String readUTF8String(Path path) throws IOException {
-        return Files.readString(path, StandardCharsets.UTF_8);
-    }
-
-    public static void writeUTF8String(Path path, String content) throws IOException {
-        Files.writeString(path, content, StandardCharsets.UTF_8);
-    }
-
+public class FileMutatorUtils {
     /**
      * The `File.deleteOnExit` method is not suited for long-running processes as it never removes the added files,
      * thus leading to a memory leak.
@@ -126,22 +106,6 @@ public class FileUtils {
     }
 
     /**
-     * <b>Blocking</b>; Waits until the specified path exists, polling every 100ms up to the given timeout.
-     *
-     * @throws InterruptedException if the thread is interrupted while waiting.
-     */
-    public static void waitUntilFileExists(Path path,
-                                           long timeoutMillis) throws InterruptedException, TimeoutException {
-        long start = System.currentTimeMillis();
-        while (!Files.exists(path)) {
-            if (System.currentTimeMillis() - start > timeoutMillis) {
-                throw new TimeoutException("File did not exist after " + timeoutMillis + " ms: " + path.toAbsolutePath());
-            }
-            Thread.sleep(100);
-        }
-    }
-
-    /**
      * Create a temporary directory that is automatically deleted on JVM exit.
      *
      * @return The path to the created temporary directory
@@ -185,24 +149,6 @@ public class FileUtils {
     }
 
     /**
-     * List all files (not directories) in the given directory up to the specified depth.
-     *
-     * @param path  The directory to list files from
-     * @param depth The maximum depth to traverse
-     * @return A set of file names (not paths) in the directory
-     * @throws IOException if an I/O error occurs
-     */
-    public static Set<String> listFilesInDirectory(Path path, int depth) throws IOException {
-        try (Stream<Path> stream = Files.walk(path, depth)) {
-            return stream
-                    .filter(p -> !Files.isDirectory(p))
-                    .map(Path::getFileName)
-                    .map(Path::toString)
-                    .collect(Collectors.toSet());
-        }
-    }
-
-    /**
      * Write the given content to the specified file, creating the file if it does not exist.
      * Uses the platform default charset.
      *
@@ -219,41 +165,8 @@ public class FileUtils {
         }
     }
 
-    public static Optional<String> readFromFileIfPresent(Path path) {
-        try {
-            return Optional.of(Files.readString(path));
-        } catch (IOException e) {
-            return Optional.empty();
-        }
-    }
-
-    public static String readStringFromResource(String resourceName) throws IOException {
-        try (Scanner scanner = new Scanner(getResourceAsStream(resourceName))) {
-            return readFromScanner(scanner);
-        }
-    }
-
-    public static String readFromScanner(Scanner scanner) {
-        StringBuilder sb = new StringBuilder();
-        while (scanner.hasNextLine()) {
-            sb.append(scanner.nextLine());
-            if (scanner.hasNextLine()) {
-                sb.append(System.lineSeparator());
-            }
-        }
-        return sb.toString();
-    }
-
-    public static InputStream getResourceAsStream(String resourceName) throws IOException {
-        InputStream resource = FileUtils.class.getClassLoader().getResourceAsStream(resourceName);
-        if (resource == null) {
-            throw new IOException("Could not load " + resourceName);
-        }
-        return resource;
-    }
-
     public static void resourceToFile(String resourceName, Path outputPath) throws IOException {
-        try (InputStream resource = getResourceAsStream(resourceName)) {
+        try (InputStream resource = FileReaderUtils.getResourceAsStream(resourceName)) {
             Files.deleteIfExists(outputPath);
             Files.copy(resource, outputPath, StandardCopyOption.REPLACE_EXISTING);
         }
@@ -265,38 +178,6 @@ public class FileUtils {
 
     public static void copy(InputStream inputStream, OutputStream outputStream) throws IOException {
         inputStream.transferTo(outputStream);
-    }
-
-    public static Set<String> listRegularFiles(Path dirPath) {
-        if (Files.notExists(dirPath)) {
-            return new HashSet<>();
-        }
-        try (Stream<Path> stream = Files.list(dirPath)) {
-            return stream
-                    .filter(Files::isRegularFile)
-                    .map(Path::getFileName)
-                    .map(Path::toString)
-                    .collect(Collectors.toSet());
-        } catch (IOException e) {
-            log.error(e.toString(), e);
-            return new HashSet<>();
-        }
-    }
-
-    public static Set<String> listDirectories(Path dirPath) {
-        if (Files.notExists(dirPath)) {
-            return new HashSet<>();
-        }
-        try (Stream<Path> stream = Files.list(dirPath)) {
-            return stream
-                    .filter(Files::isDirectory)
-                    .map(Path::getFileName)
-                    .map(Path::toString)
-                    .collect(Collectors.toSet());
-        } catch (IOException e) {
-            log.error(e.toString(), e);
-            return new HashSet<>();
-        }
     }
 
     /**
@@ -326,41 +207,6 @@ public class FileUtils {
             Path targetPath = dirPath.resolve(backupFolderName).resolve(newFileName);
             Files.move(storageFilePath, targetPath, StandardCopyOption.REPLACE_EXISTING);
         }
-    }
-
-    public static HttpURLConnection downloadFile(URL url,
-                                                 Path destinationPath,
-                                                 Observable<Double> progress) throws IOException {
-        HttpURLConnection connection = (HttpURLConnection) url.openConnection();
-        int fileSize;
-        try {
-            connection.connect();
-            try (InputStream inputStream = new BufferedInputStream(connection.getInputStream());
-                 OutputStream outputStream = Files.newOutputStream(destinationPath)) {
-                // If server does not provide contentLength it is -1
-                fileSize = connection.getContentLength();
-                double totalReadBytes = 0d;
-                byte[] buffer = new byte[1024];
-                int bytesRead;
-                while ((bytesRead = inputStream.read(buffer)) != -1) {
-                    outputStream.write(buffer, 0, bytesRead);
-                    if (fileSize != -1) {
-                        totalReadBytes += bytesRead;
-                        progress.set(totalReadBytes / fileSize);
-                    }
-                }
-                if (fileSize == -1) {
-                    progress.set(1d);
-                }
-            }
-        } finally {
-            connection.disconnect();
-        }
-        return connection;
-    }
-
-    public static boolean hasResourceFile(String fileName) {
-        return FileUtils.class.getClassLoader().getResource(fileName) != null;
     }
 
     public static void copyDirectory(Path sourceDirPath,
@@ -395,53 +241,6 @@ public class FileUtils {
         }
     }
 
-    public static Set<String> listResources(String resourceDir) throws IOException, URISyntaxException {
-        // Resource paths always use forward slashes on all OS.
-        String normalized = resourceDir.replace('\\', '/');
-        if (!normalized.endsWith("/")) {
-            normalized = normalized + "/";
-        }
-        URL dirURL = FileUtils.class.getClassLoader().getResource(normalized);
-        if (dirURL == null) {
-            throw new IOException("Resource directory not found: " + normalized);
-        }
-        String protocol = dirURL.getProtocol();
-        if ("file".equals(protocol)) {
-            Path dirPath = Path.of(dirURL.toURI());
-            if (!Files.isDirectory(dirPath)) {
-                throw new IOException("Resource path is not a directory: " + dirPath);
-            }
-            try (Stream<Path> stream = Files.walk(dirPath)) {
-                return stream
-                        .filter(Files::isRegularFile)
-                        .map(path -> dirPath.relativize(path).toString().replace(File.separatorChar, '/'))
-                        .collect(Collectors.toSet());
-            }
-        } else if ("jar".equals(protocol)) {
-            String jarUrlPath = dirURL.getPath();
-            int bangIndex = jarUrlPath.indexOf('!');
-            String jarFilePath = jarUrlPath.substring(0, bangIndex);
-            if (jarFilePath.startsWith("file:")) {
-                jarFilePath = jarFilePath.substring(5);
-            }
-            jarFilePath = URLDecoder.decode(jarFilePath, StandardCharsets.UTF_8);
-            Set<String> resourceFiles = new HashSet<>();
-            try (JarFile jar = new JarFile(jarFilePath)) {
-                Enumeration<JarEntry> entries = jar.entries();
-                while (entries.hasMoreElements()) {
-                    JarEntry entry = entries.nextElement();
-                    String name = entry.getName();
-                    if (!entry.isDirectory() && name.startsWith(normalized)) {
-                        resourceFiles.add(name.substring(normalized.length()));
-                    }
-                }
-            }
-            return resourceFiles;
-        } else {
-            throw new IOException("Unsupported URL protocol: " + protocol);
-        }
-    }
-
     public static void copyResourceDirectory(String resourceDir,
                                              Path path) throws IOException, URISyntaxException {
         // Resource paths always use forward slashes on all OS.
@@ -449,7 +248,7 @@ public class FileUtils {
         if (!normalized.endsWith("/")) {
             normalized = normalized + "/";
         }
-        Set<String> resources = listResources(normalized);
+        Set<String> resources = FileReaderUtils.listResources(normalized);
         for (String resourceFile : resources) {
             String resourcePath = normalized + resourceFile; // classpath uses '/'
             Path targetFilePath = path.resolve(resourceFile); // preserve relative layout
@@ -458,11 +257,42 @@ public class FileUtils {
                 if (parentPath != null) {
                     Files.createDirectories(parentPath);
                 }
-                FileUtils.resourceToFile(resourcePath, targetFilePath);
+                resourceToFile(resourcePath, targetFilePath);
             } catch (IOException e) {
                 log.error("Could not copy resource {} to {}", resourcePath, targetFilePath, e);
                 throw e;
             }
         }
+    }
+
+    public static HttpURLConnection downloadFile(URL url,
+                                                 Path destinationPath,
+                                                 Observable<Double> progress) throws IOException {
+        HttpURLConnection connection = (HttpURLConnection) url.openConnection();
+        int fileSize;
+        try {
+            connection.connect();
+            try (InputStream inputStream = new BufferedInputStream(connection.getInputStream());
+                 OutputStream outputStream = Files.newOutputStream(destinationPath)) {
+                // If server does not provide contentLength it is -1
+                fileSize = connection.getContentLength();
+                double totalReadBytes = 0d;
+                byte[] buffer = new byte[1024];
+                int bytesRead;
+                while ((bytesRead = inputStream.read(buffer)) != -1) {
+                    outputStream.write(buffer, 0, bytesRead);
+                    if (fileSize != -1) {
+                        totalReadBytes += bytesRead;
+                        progress.set(totalReadBytes / fileSize);
+                    }
+                }
+                if (fileSize == -1) {
+                    progress.set(1d);
+                }
+            }
+        } finally {
+            connection.disconnect();
+        }
+        return connection;
     }
 }
