+++ http-api/src/main/java/bisq/http_api/ApiTorOnionService.java
@@ -18,7 +18,7 @@
 package bisq.http_api;
 
 import bisq.common.application.Service;
-import bisq.common.file.FileUtils;
+import bisq.common.file.FileMutatorUtils;
 import bisq.common.network.TransportType;
 import bisq.network.NetworkService;
 import bisq.security.SecurityService;
@@ -68,7 +68,7 @@ public class ApiTorOnionService implements Service {
                     log.info("{} published for {}", onionAddressWithPort, identityTag);
                     try {
                         Path path = appDataDirPath.resolve(identityTag + "_onionAddress.txt");
-                        FileUtils.writeToPath(onionAddressWithPort, path);
+                        FileMutatorUtils.writeToPath(onionAddressWithPort, path);
                         return true;
                     } catch (IOException e) {
                         log.error("Error at write onionAddress", e);
