+++ network/tor/tor/src/main/java/bisq/network/tor/process/control_port/ControlPortFileParser.java
@@ -17,7 +17,7 @@
 
 package bisq.network.tor.process.control_port;
 
-import bisq.common.file.FileUtils;
+import bisq.common.file.FileReaderUtils;
 
 import java.io.IOException;
 import java.nio.file.Path;
@@ -27,7 +27,7 @@ public class ControlPortFileParser {
 
     public static int parse(Path controlPortFilePath) {
         try {
-            String fileContent = FileUtils.readUTF8String(controlPortFilePath);
+            String fileContent = FileReaderUtils.readUTF8String(controlPortFilePath);
             if (isControlPortFileReady(fileContent)) {
                 for (String line : fileContent.split("\n")) {
                     // Lines end on Windows with "\r\n". Previous String.split("\n") removed "\n" already.
